+++ projects/ng-halfmoon/src/lib/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './applier';
 export * from './sizing';
+export * from './appearance';